Code refactored, tests: extract shared KV-cache eval helpers into kv_cache_{eval,bench}_common
