+++ tests/test-kv-cache-longbench.py
@@ -33,8 +33,6 @@ import logging
 import os
 import subprocess
 import sys
-import threading
-import time
 from concurrent.futures import ThreadPoolExecutor, as_completed
 from dataclasses import dataclass, field
 from datetime import datetime
@@ -46,22 +44,13 @@ logger = logging.getLogger(__name__)
 SCRIPT_DIR = Path(__file__).resolve().parent
 LONGBENCH_BENCH = SCRIPT_DIR / "longbench-bench.sh"
 
-# ── Bits-per-weight for KV cache types ──────────────────────────────────────
-
-BPW = {
-    "f16":    16.0,
-    "q8_0":   8.5,
-    "q4_0":   4.5,
-    "tbq3_0": 4.25,
-    "tbq4_0": 5.25,
-    "pq3_0":  3.25,
-    "pq4_0":  4.25,
-}
-
-
-def bpw_label(k_type: str, v_type: str) -> str:
-    return f"K:{BPW.get(k_type, '?')} V:{BPW.get(v_type, '?')}"
-
+sys.path.insert(0, str(SCRIPT_DIR))
+import kv_cache_eval_common as kv_common
+from kv_cache_eval_common import (
+    BPW, bpw_label, ModelDef, MODELS_DEFAULT,
+    ProgressTracker, _print_lock, _fmt_duration,
+    _next_available_dir, _find_latest_dir,
+)
 
 # ── Quant configurations ───────────────────────────────────────────────────
 # Same shape as test-kv-cache-ruler.py so cross-bench comparisons line up.
@@ -87,39 +76,6 @@ QUANT_CONFIGS_SUBSET = [
     ("tbq3_0", "pq3_0"),
     ("tbq4_0", "q4_0"),
     ("q4_0",   "q4_0"),
-]
-
-# ── Model definitions ──────────────────────────────────────────────────────
-
-
-@dataclass
-class ModelDef:
-    path: str
-    tokenizer: str
-    label: str
-    family: str  # "Llama-3.1-8B", "Ministral-8B" — used for cross-model aggregation.
-
-
-MODELS_DEFAULT = [
-    # Primary: blog chart model.
-    ModelDef(
-        path="models/Meta-Llama-3.1-8B-Instruct-Q4_K_M.gguf",
-        # Un-gated mirror — same tokenizer.json SHA as the official meta-llama
-        # repo, no HF login required.
-        tokenizer="NousResearch/Meta-Llama-3.1-8B-Instruct",
-        label="Llama-3.1-8B",
-        family="Llama-3.1-8B",
-    ),
-    # Secondary: paper's other model (§4.3). Ministral-8B (Mistral AI's 2024
-    # release) is the closest GGUF match — paper text says "Ministral-7B-Instruct"
-    # but Mistral AI only released the 8B-Instruct-2410 variant; community
-    # consensus is that that's what the paper actually evaluated.
-    ModelDef(
-        path="models/Ministral-8B-Instruct-2410-Q4_K_M.gguf",
-        tokenizer="mistralai/Ministral-8B-Instruct-2410",
-        label="Ministral-8B",
-        family="Ministral-8B",
-    ),
 ]
 
 # ── Config presets ─────────────────────────────────────────────────────────
@@ -173,39 +129,6 @@ class Job:
     csv_file: str = ""
 
 
-# ── Progress tracking ──────────────────────────────────────────────────────
-
-_print_lock = threading.Lock()
-
-
-class ProgressTracker:
-    def __init__(self, total: int):
-        self.total = total
-        self.completed = 0
-        self.start_time = time.time()
-        self.lock = threading.Lock()
-
-    def tick(self) -> str:
-        with self.lock:
-            self.completed += 1
-            elapsed = time.time() - self.start_time
-            avg = elapsed / self.completed if self.completed > 0 else 0
-            remaining = max(0, self.total - self.completed)
-            eta = _fmt_duration(avg * remaining) if self.completed > 0 else "?"
-            return f"[{self.completed}/{self.total}] elapsed {_fmt_duration(elapsed)}, ETA {eta}"
-
-
-def _fmt_duration(secs: float) -> str:
-    m, s = divmod(int(secs), 60)
-    h, m = divmod(m, 60)
-    if h > 0:
-        return f"{h}h{m:02d}m"
-    return f"{m}m{s:02d}s"
-
-
-_progress: Optional[ProgressTracker] = None
-
-
 # ── Run a single longbench-bench.sh job ────────────────────────────────────
 
 
@@ -239,48 +162,16 @@ def run_job(job: Job, extra_args: list[str], output_dir: Path) -> Optional[str]:
     cmd.extend(extra_args)
 
     label = f"[GPU{job.gpu_id}] {job.model.label} K={job.k_type} V={job.v_type}"
-    cmd_str = " ".join(cmd)
     log_path = csv_path.replace(".csv", ".txt")
-    with _print_lock:
-        logger.info("  START: %s", label)
-        logger.info("  $ %s", cmd_str)
-        logger.info("  log: %s", log_path)
-        logger.info("")
-
-    try:
-        # 12 h cap covers the paper preset (~24h * 0.5 if we get any GPU
-        # speedup); reraise as a soft fail rather than blocking everything.
-        # `errors="replace"` defends against malformed UTF-8 in the runner's
-        # captured stdout — `printf "%.Ns"` in bash truncates at *byte*
-        # boundaries and can split a multi-byte codepoint mid-sequence
-        # (we also fix this on the bash side, but keep the Python safety
-        # net so a single noisy prediction never crashes the whole run).
-        result = subprocess.run(
-            cmd, env=env, capture_output=True, text=True,
-            errors="replace", timeout=12 * 3600,
-        )
-        if result.returncode != 0:
-            with _print_lock:
-                logger.info("  FAIL:  %s", label)
-                logger.info("         stderr: %s", result.stderr[-500:])
-            return None
-    except subprocess.TimeoutExpired:
-        with _print_lock:
-            logger.info("  TIMEOUT: %s", label)
-        return None
-
-    # The runner prints "Category-avg score: 50.06%" — parse it for the
-    # short DONE log line. The real per-task data goes via the CSV.
-    category_avg = "?"
-    for line in result.stdout.splitlines():
-        if "Category-avg score:" in line:
-            category_avg = line.strip().split("Category-avg score:")[1].strip()
-            break
-
-    progress_str = _progress.tick() if _progress else ""
-    with _print_lock:
-        logger.info("  DONE:  %s  =>  %s  %s", label, category_avg, progress_str)
-    return csv_path
+    # 12 h cap covers the paper preset (~24h * 0.5 if we get any GPU
+    # speedup); reraise as a soft fail rather than blocking everything.
+    ok = kv_common.run_bench_subprocess(
+        cmd, env, label, log_path,
+        timeout=12 * 3600,
+        score_marker="Category-avg score:",
+        errors="replace",
+    )
+    return csv_path if ok else None
 
 
 # ── Collect all CSV results ────────────────────────────────────────────────
@@ -447,32 +338,6 @@ def write_combined_csv(results: list[CellResult], path: str):
 # ── Output dir helpers (RULER pattern) ─────────────────────────────────────
 
 
-def _next_available_dir(base: Path) -> Path:
-    if not base.exists() or not any(base.iterdir()):
-        return base
-    i = 1
-    while True:
-        candidate = base.parent / f"{base.name}{i}"
-        if not candidate.exists() or not any(candidate.iterdir()):
-            return candidate
-        i += 1
-
-
-def _find_latest_dir(base: Path) -> Optional[Path]:
-    latest = None
-    if base.exists() and any(base.iterdir()):
-        latest = base
-    i = 1
-    while True:
-        candidate = base.parent / f"{base.name}{i}"
-        if candidate.exists() and any(candidate.iterdir()):
-            latest = candidate
-            i += 1
-        else:
-            break
-    return latest
-
-
 def _job_csv_path(output_dir: Path, job: Job) -> Path:
     name = f"{job.model.label}_{job.k_type}_{job.v_type}"
     name = name.replace("/", "_").replace(" ", "_")
@@ -611,8 +476,7 @@ def main():
     csv_paths: list[str] = list(skipped_csvs)
 
     if all_jobs:
-        global _progress
-        _progress = ProgressTracker(len(all_jobs))
+        kv_common.set_progress(ProgressTracker(len(all_jobs)))
         logger.info("\nRunning %d jobs across %d GPU(s)...\n",
                     len(all_jobs), len(gpu_ids))
         with ThreadPoolExecutor(max_workers=len(gpu_ids)) as pool:
